Aqueduct: Allow multiple synthesize parents and remove DependencyContainerRegistry (#8593)
This change is a preceding change needed for #8074. Removing IFluidObject from aqueduct, specifically synthesize will change much of the typing, as it is heavily dependent on keyof IFluidObject. This change relates, as in making that change found some problems with DependencyContainerRegistry that this change fixes:

The typing of DependencyContainerRegistry was incorrect, as it allowed undefined returns which are not allowed or expected via the rest of synthesize typeing
DependencyContainerRegistry was coupled to the class, rather than the IFluidDependencySynthesizer interface
DependencyContainerRegistry couldn't be made generic to move away from keyof IFluidObject
To fix the above issues, i've moved away from DependencyContainerRegistry and instead enabled DependencyContainer to support multiple parents, so uses can simply pass in a DependencyContainer directly, and that is added, along with scope to the child registry. Additionally, i've deprecated a number of fields on IFluidDependencySynthesizer, and these will now only exist on DependencyContainer itself