Fix on-demand summary test failure with real service
Modify the test to make sure the ack/nack of the final on-demand summary occurs before attempting to stop the summarizer. Otherwise, the summarizer won't stop (it will be waiting for the lock to be released), and an assert in the test will fire when we let a summary go ahead after the stop call returns.