Improve Bundle Size Tests (#8603)
In #8599 i didn't see the bundle size impact i had expected. There are a couple reasons that could be, but I also noticed our bundle coverage had some gaps, it didn't cover the loader, or aqueduct, so this change adds that coverage, and renames container to containerRuntime for clarity. The aqueduct test should give a more representative bundle of multiple packages