(split from Whiteboard) Merged PR 37075: build: updates package dependencies (except typescript, colors)
Updates to package dependencies from `npm run update-package-dependencies`.

## colors
The maintainer of the colors package has introduced a bug (on purpose) in version 1.4.1. This PR pins the package to version 1.4.0.

## typescript
Following the pattern of the last three dependency updates PR (see last here: https://dev.azure.com/intentional/intent/_git/typescript-pipe/pullrequest/36923 ), we're not updating typescript to 4.5.4 because it introduces a compilation error. That error is being resolved in a different PR.

Related work items: #70416