Remove IFluidObject from Aqueduct (#8891)
IFluidObject is deprecated. In this release we have removed all IFluidObject from the aqueduct package.
This impacts the following public apis:

getDefaultObjectFromContainer
getObjectWithIdFromContainer
getObjectFromContainer
PureDataObject.getFluidObjectFromDirectory
ContainerServiceRegistryEntries
SingletonContainerServiceFactory.getService
In general the impact of these changes should be transparent. If you see compile errors related to Fluid object provider types with the above apis, you should transition those usages to FluidObject which is the replacement for the deprecated IFluidObject.

fixes #8074

Additional we are able to remove the internal utility type DataObjectType now that we have better handling for undefined our our generic parameters