Cleanup some duplicate telemetry helpers (#9016)
extractLogSafeErrorProperties:

This helper existed in both telemetry-utils and common-utils packages. The copy in common-utils was out of date and unused in this repo, and has now been removed. Tests were moved to telemetry-utils and updated.

extractSafePropertiesFromMessage / extractLogSafeMessageProperties:

extractLogSafeMessageProperties was a non-exported copy (almost!) of extractSafePropertiesFromMessage. So I removed it and updated callers to use the shared one. The prop names added by the shared one are all prefixed with message, but this was not the case for the duplicate. I will update #9014 with the changes in property names for the two impacted codepaths once this merges.