Do not close container from reconnectOnError even if canRetry is false (#8964)
Fixes #8570

There should be no behavior change with this PR. But the flow/expectations are much clearer.

Previously, 2/3 codepaths entering reconnectOnError would stomp over canRetry with true, to avoid closing the container due to hiccup on the socket (assumption being, any fatal error state will cause container to close on reconnect attempt). Now, rather than overwriting canRetry, I've let that stay as-is and updated reconnectOnError to simply log if canRetry is false.

The third codepath entering reconnectOnError (nack) has a case where canRetry is false, so in that case rather than leaning on reconnectOnError to close the container, I check canRetry first and close right away (then reconnectOnError will return early).

Note: My previous analysis that stated that documentDeltaConnection's disconnect event could lead to this code path if raised during odspDocumentDeltaConnection.initialize was incorrect - at that point in time, connectionManager hasn't registered event listeners, of course, since the connection object is never successfully created/returned.

With this in mind (and given the refactor for the non-retryable nack), I don't actually expect the new log event "reconnectingDespiteFatalError" to ever be hit. And even if it is, it's an expected/supported code path so I'm logging it as a generic event, not an error event.