Add externalErrorToFluidError as optional on normalizeError
In other words, promote semantics of DPE.wrapIfUnrecognized
to normalizeError itself.
This makes sense - it lets code pass through our errors
and classify external errors in a consistent way.
Will also be used in connectCore.
Foreshadowing of when our errors are returned via API
and any caught error is external.