Rename telemetry prop (#9045) (#9172)
The office telemetry logging code doesn't like dashes in telemetry prop names. I thought I fixed this in #8449 but the fix was incomplete. That PR addressed the error case, but I didn't realize that for successful requests we were stashing the same telemetry prop on the internal response object and later logging it as well.

So I fixed that, and renamed the property bag to be more clear that it's for logging and not for business logic.

Co-authored-by: Mark Fields <markfields@users.noreply.github.com>