(split from Whiteboard) Merged PR 38232: feat: BREAKING CHANGE: Make BuildNode identifier optional
A `BuildNode` may now omit its identifier property, in which case the `SharedTree` will choose an identifier automatically for the new node. This changes the public API for building nodes in a non-breaking fashion, but does not change the underlying op and summary format, which continues to use nodes that require IDs (`BuildNodeInternal`, `ChangeNode`).

This PR also includes various related cleanup and refactors:
* Public change types are now located in ChangeTypes.ts
* The `Side` enum is now in the generic folder
* `SharedTree` now implements the new `IdGenerator` interface; its `generateId` method has been renamed to `generateNodeId`
* `ConstraintEffect` is now in PersistedTypes.ts
* `convertTreeNodes` is now looser - it delegates transformation of each node to the caller and does the minimal work to place new nodes under the proper traits, but no longer copies known properties (`identifier`, `definition`, `payload`).

Note: There are many tests that can now be simplified since they do not need to provide identifiers for every build node. Those will be updated in a future PR.

Related work items: #72480