(split from Whiteboard) Merged PR 38085: feat: Adds WaitableDocument interface, exposes waitForUnsentLocalEdits on document
This change is part of the offline work support effort. Right now, some edits (like adding a template) will queue up in shared-tree and asynchronously find their way to the fluid container. When the board is being closed, we want to save all of the fluid container's pending local state to indexeddb. The problem, though, is that because some edits can be asynchronously handled in the fluid document, we could potentially lose information.

This PR resolves that issue by exposing a new interface, `WaitableDocument`, which exposes a `waitForUnsentLocalEdits` function. This allows whiteboard to wait for unsent local edits to be passed off to the fluid container, which we can then save off from the pendingLocalState of the fluid container to save and use later.

Related work items: #68613