Cleanup describeCompat and enable skip, only, and noCompat (#9222)
Was debugging some end to end test, and wanted to use skip and only in on our compact tests. I also added an option to easily disable compat. in the future we could also extend skip to support drivers or compat versions for more targeted skipping.

To make things easier to read i put itExpects and describeCompat in their own files as well