(split from Whiteboard) Merged PR 38454: improvement: BREAKING CHANGE: convert node IDs in `validateStablePlace` and `validateStableRange`
These functions now attempt to convert node IDs and will fail appropriately if the conversions are impossible. If they succeed, they will return the converted form of the input so that callers don't need to repeat the ID conversion in the event that they need the converted IDs.

Related work items: #70358