Add enum support in API validator (#8901)
Fixes #8029

Add support for examining enums in the API validator along with tests for the cases where an existing enum might change. This PR also includes the commits from #8900 which will be merged first separately (to see only the enum changes before that happens, exclude the first commit). Unlike with classes, it is easier to check enums entirely ourselves rather than trying to leverage the ts compiler.