Add virtual patch support to build-tools (#9297)
Fixes #9292

Add support for virtual patch versioning (where "minor" changes are a 1000 increment to the patch component, which reserves space for actual patches) in build-tools. This is enabled through a --virtualPatch flag when running bump-version and will be used during the transition period between the 0.58 release and 1.0.

Q: What is virtual patch
A: The current beta versioning scheme we use (0.x.x) does not have room to differentiate major/minor/patch because we only have 2 version components. We can simulate this by making the second component represent the major version, and combine minor and patch into the third by representing minor as a 1000 increment and patch as a 1 increment. This reserves number space (999 of them) between each minor version, allowing room for patches.