(split from Whiteboard) Merged PR 38628: test: Verify version upgrade succeeds for tree initialized with older writeFormat
Existing update tests generally observed what happened to trees initialized with a `writeSummaryFormat` matching the version being upgraded to. This confirms that a tree initialized with an older write summary format can also update, as long as the newer format is known to it. This is a practically useful scenario for being able to do partial rollouts of features, e.g. by ECS.

Related work items: #73073