MergeTree: Unify removed client id and overlap clients (#9369)
This change unifies the removedClientId and removedClientOverlap into a single array, removedClientIds.

This simplifies our handling in a number of places, as we no longer need to check and manage both. In fact, there were a couple bugs where both were not handled appropriately.

In addition, I've improved how we handle IRemovalInfo such that it is now all or nothing, which removes the need for a number of unsafe non-null assertions, and make the code simpler by removing the need to repeated non-null checks.

this may fix #9138, but it hard to know as we don't have much telemetry around the issue.