(split from Whiteboard) Merged PR 38338: docs: Merged changes from master
Automatic merge from the `master` branch to the `wiki` branch failed. This is likely due to a merge conflict. Please resolve any issues by checking out wiki-pipeline/master-to-wiki and making any necessary changes.
Note - If you see changes outside the docs/ folder that are conflicting, this is unexpected behavior and could lead to future pipeline issues. The only way this should happen is if there are changes outside the docs/ folder on the `wiki` branch which should not have been made. Please revert those changes on that branch, and make them on the `master` branch instead.
Once you have finished, please get approval for this PR and complete it with the following options
* Use "Merge (no fast forward)" as opposed to "Squash".
* Delete the branch (wiki-pipeline/master-to-wiki) when you complete the PR - the wiki pipeline uses the existence of the branch to determine whether or not a new pipeline run can be executed. If the branch is not deleted, the wiki pipeline will not run on new changes.
Any changes made to wiki-pipeline/master-to-wiki and then merged into the `wiki` branch should automatically make their way back to the `master` branch via the wikiMaster -> master pipeline, assuming it is also running.

Related work items: #68613, #72678