Cherry pick bump to 0.58.1000 (and related bumps) from main to release/0.58 (#9428)
One persistent point of confusion around the new versioning numbers has been what something like 0.58.1 means (if it is a patch or a minor like with the old scheme). When formatted as 0.58.0001 it becomes much clearer that it represents a patch, but existing tools like npm and lerna do not consider this sort of padding as valid - attempting to set the version to such with either tool will cause it to fail. We can instead start versioning at 0.58.1000 instead of 0.58.0 to force 4 digits. This is easier to explain and understand than the 0.58.1 ambiguity.

main will then be bumped to 0.58.2000. Related changes to tooling will come later so as not to block the release.