Bump packages that will be released in release/0.58.0 again to account for adjusted versioning scheme (#9429)
See #9428 for matching change.

To address the ambiguity with something like 0.58.1 in the new versioning scheme, we are moving to 1 indexing for minor releases to retain 4 significant digits because padding 0s are invalid. E.g. 0.58.1000 will be the first minor release instead of 0.58.0. main needs to be bumped to 0.58.2000 as a result.