Fix virtualPatch option for branch command in bump-version tool (#9380)
The branch command was short-circuiting when parsing options and not picking up the virtualPatch option. Remove the short-circuit so additional options may be parsed. There is already a later check for conflicting options that will exit if branch commands pick up additional unwanted options as a result of this change.