Update bump-version tool to use 1 indexing for virtual patch (#9459)
Update the bump-version tooling to use 1 indexing for the minor component when bumping the major component for virtual patches. This preserves 4 digits for the last version component (that holds minor and patch information) for all versions, and skips over the x.x.0 version. This avoids confusion around x.x.1 and x.x.1000 due to the differences in number of digits. 0 padding is not an option because it is not considered valid semver.