FI main to next (#9515)
* Move internal implementation of the odsp driver cache to the Fluid Framework Repository (#9418)

* initial commit

* Update tsconfigs

* Compile fixes

* Make tests pass

* Add readme

* Update formatting

* Some eslint suppressions and fixes

* Fix formatting lint issues

* Disable some lint rules

* More disables

* Fix last of eslint issues

* Policy check fixes

* Policy update

* Pin url-parse types

* Add comments

* Undo lerna package changes

* Update lockfile

* Undo content changes

* Undo content changes

* Undo content

* Remove deprecated API

* Rename folder

* Autogenerated files

* Remove more internal references

* Update packages/drivers/driver-web-cache/package.json

Co-authored-by: Helio Liu <59622401+heliocliu@users.noreply.github.com>

* Update packages/drivers/driver-web-cache/package.json

Co-authored-by: Helio Liu <59622401+heliocliu@users.noreply.github.com>

* Minor cleanups

Co-authored-by: Helio Liu <59622401+heliocliu@users.noreply.github.com>

* Update bump-version tool to use 1 indexing for virtual patch (#9459)

Update the bump-version tooling to use 1 indexing for the minor component when bumping the major component for virtual patches. This preserves 4 digits for the last version component (that holds minor and patch information) for all versions, and skips over the x.x.0 version. This avoids confusion around x.x.1 and x.x.1000 due to the differences in number of digits. 0 padding is not an option because it is not considered valid semver.

* Update BREAKING.md format to include upcoming change notes (#9457)

Update the formatting for 0.58 section and the instructions to include upcoming change notes. This is to be more explicit about upcoming breaking changes to help consumers with planning them, because currently they tend to get lost with the other breaking entries. This is also to help build habits around announcing upcoming breaking changes instead of dumping them on consumers only when they actually break.

* Update tinylicious and gitrest, and use release packages everywhere (#9473)

* Use strict eslint config in fluid-framework (#9476)

* Add recursive firing of dispose events of sub directory deletion (#9451)

* Remove outdated socket.io client types package (#9479)

* Fixes the FlushMode validation in PendingStateManager which expected Immediate to be the default FlushMode (#9432)

* Organize and document summarizer telemetry (#9278)

Add typing and comments for fields logged in summarizer telemetry. Fix a double increment of RunningSummarizer.totalSuccessfulAttempts.

* Bump eslint version for main, and bump used versions of eslint after release (#9493)

* Update to latest tinylicious throughout repo (#9494)

* gitrest: pre-compute latest summary for Historian caching (#9490)

* gitrest: add FS Manager interface (#9496)

* r11s-driver: handle node-fetch network errors (#9498)

* Remove GC from odspCache implementation. (#9415)

* Remove GC from odspCache implementation by using a local Expiration cache with a single entry per
documentId. Also, taking the chance to remove 2 secs timeouts from the UTs and start relying on the cleanup at the end of the test to remove the cache entries just like the actual implementation that implements from the EpochTracker does.

* Pass ISequencedDocumentMessage to interval creation API (#9465)

Pass msg to interval creation API so we attach to the correct text segment on remote op

* Remove duplicate nesting of server packages from lerna-package-lock.json (#9507)

Co-authored-by: Christian Gonzalez <1581488+christiango@users.noreply.github.com>
Co-authored-by: Matt Rakow <ChumpChief@users.noreply.github.com>
Co-authored-by: Tyler Butler <tyler@tylerbutler.com>
Co-authored-by: Jatin Garg <48029724+jatgarg@users.noreply.github.com>
Co-authored-by: Navin Agarwal <45832642+agarwal-navin@users.noreply.github.com>
Co-authored-by: Paul Leathers <pleath@users.noreply.github.com>
Co-authored-by: Zach Newton <znewton@microsoft.com>
Co-authored-by: NicholasCouri <nichoc@microsoft.com>