Snapshot Tests: Strict Channels and Non-comparable Channels (#9441)
The pervious implementation of Unknown Channels had a major issue where we were losing data simply because the data was non-deterministic.

This change does a few things. It allows adding channel factories that are non-deterministic, like shared matrix, but enabled excluding them from comparison. This allows use to keep the data, which will still be tested in hydration.

Additionally, this adds a mode, strictChannels to the replayTool. This is only set in the snapshot tests, and will cause failure if a channel type is not found. This will prevent the addition of new snapshots to the snapshot tests which include channels we don't have explicit handling for. This will prevent further data loss.

TestData PR: microsoft/FluidFrameworkTestData#40

related #6957 #9405