Remove some peer dependencies (#9563)
Using peer dependencies with our build tooling is unneeded because the tools are intended only for use in our repo, and peer dependencies are a hassle for downstream consumers.

Note that we have legitimate peer dependencies declared in the azure-client and Property DDS packages. Those dependencies are not addressed in this PR. See #7913 for more details.