Ensure all packages depend on common-definitions 20.1 (#8624)
Months ago I made the mistake of introducing a breaking change between 20.0 and 20.1, and client consumes 0.20.1. So it's best to get everyone on the same page here. Anyway, due to semver and npm install behavior, I'm pretty certain that anyone consuming client will be installing 0.20.1 so this shouldn't cause any changes as it rolls out.

Specifically, this causes breaks in some cases when trying to do --symlink:full.