Enable type validation for all client packages (#9673)
This change uses a new improved version of the type validation tool to generate validation tests for all packages (#9674).

Generally this change in all done via automation, but there are a couple other changes.

Remove type validation from existing packages, and move to new model
Fix fluid-build non-conformant packages
All automatic type validation generation is removed. It is to slow to run on every build. It should only need to be run after release, and in the presence of new breaking changes that must be suppressed