Updates to bump-version tool around policy-check, branch command, getting release versions (#9662)
Sort of addresses #8946, parts of #8948

This PR does 3 things:

Update the policy-check step during release to output a warning when not run on main. This is primarily motivated by tagging behavior, where tags should only be generated in main to avoid unexpected tag aliasing.
Change the branch command in the bump-version tool to releaseBump which better matches its usage. Also adjust its behavior to do the expected thing depending on which branch is running it (because main and next expect different levels of bumps)
Add a writeReleaseVersions command to bump-version which writes out the latest versions of packages that would exist if packages in the repo were to be released in its current state. This is to automate the step for get versions for our partners.