TypeValidator: Add flatter broken format, and generateOnly mode (#9901)
Previously the type validation tool would compare against multiple previous versions. This didn't provide value and increased overhead, so we moved it to a single version. However, we still kept the broken data nested under the previous version number, this makes the merges between main and next harder, as they don't have matching versions, so this change support a new default format that does not nest the broken data under version numbers.

Additionally, i've added a generateOnly mode that only generates the test, and does not update the package.json. I want to experiment with the performance of this mode, as it may be fast enough to integrate in the build, as the need to generate independently is causing confusion.