Add Support for generating type test in fluid-build (#9914)
This change adds knowledge of fluid-type-validation to both fuild-build and the package checker. The fluid-build integration ensures we only generate the test types when the package json has changed as we only need to generate type test when either the version of the package changes, or new entries are added to the broken list.