Integrate type tests gen and fix non-conformant scripts (#9918)
This change integrates type test generation into the build. With fluid-build and build:fast type test generation will only run when the package json changes, as part of #9914. This works as we only need to regenerate when the version changes, or when the broken section changes. This adds minimal overhead to most incremental builds.