MergeTree: Fix split handling for local references (#9961)
In this change I extended the insertAtRefPos workload we used in conflict farm to randomly introduce transient, simple, and slide on remove local references.

transient references are not added to the segment or its local reference collection
simple are added to the segment, but will not survive segment delete
slide on removed will survive delete, and slide to the nearest positions
In adding this workload i hit an assert in the split handling, so i fixed that bug, and also added a regression case for that specific issue.

related to #1008