TypeValidator: Fix Disabled Gen and Removed Types Bugs (#10093)
This change fixes two issues:

Disabled packages could not be re-enabled, as if the previous version was disabled the type test generation tool would fail to load it and bail out on generating test for the current package
There was a bug in how we handled removed types. This changes the behavior to just remove the test for removed types that are listed in the broken section. This is necessary as removed types can break both the check, and the declaration of the get or use methods.