Add tests for `connect()` and `disconnect()` (#10009)
This PR adds tests for the newly added `connect()` and `disconnect()` functions (see https://github.com/microsoft/FluidFramework/pull/9439). Tests are added to `@fluidframework/test-end-to-end-tests`.

The tests first check that the connection state is changed when using `connect()`/`disconnect()`. Next, we ensure that op processing is paused after `disconnect()`, and resumes after `connect()`.

Closes #9746