Deprecate Local Reference Class and Obsolete ReferencePosition Members (#10167)
This class deprecates the LocalReference class, and removes a number obsolete reference position members.

There will be a follow up pr in main with brings in new methods for working with ReferencePosition, and another follow up in main that hides the LocalReference class itself from export. I will add a breaking description with that PR, as it will bring the replacements. Most of the deprecated methods are unused, or only used internally so there should be minimal impact to consumers.

This is necessary to keep the implemenation of LocalReference internal, so that we can bidirectionally link to the LocalReferencCollection for efficient operations on the collection itself from LocalReference. related to #1009

AB#145 AB#148