Local reference collection linked list (#10281)
Move local reference collection to track references via a linked list rather than via arrays, this is primarily for performance, as it removes the need to allocate and modify arrays frequently, which is costly. These operation will become frequent when local references are used to track removed segments for undo-redo.

related to #1009 AB#146