Fix buildVersion script to calculate the latest tag (#10372)
The build pipeline is incorrectly tagging versions from older release branches as latest. This change updates the buildVersion script in build-tools to use semver to compare the versions and determine whether the build in the pipeline should be tagged latest or not.

I also discovered that there is a separate copy of the buildVersion script, and it has received independent updates. That is, the typescript source and the javascript called by the build pipeline have diverged. This PR attempts to merge the changes from the JS back into TS in addition to the isLatest fixes.