Trying to add some tests.
I updated CatchUpMonitor to take a DM instead of Container.
This is not good because DM has to be connected
but you can't ask it if it is.

New plan: Revert all changes to waitContainerToCatchUp.
Move CatchUpMonitor into ConnectionStateHandler,
and have it take from Container
some subset of DeltaManager
which can be replaced with "immediate"
implementation when feature gate is off.

This approach is best because the lifecycle
of the CatchUpMonitor is tightly coupled
to Container's connection state/lifecycle.