WIP - telemetry on processCore and some map emit()s
Not liking the approach of intercepting in the emitter, it makes it so the DDS need to pass the sequenced message object way into their logic so they can pass the relevant infromation into the emit calls.

Also noticed that SharedSubDirectory is implemented by directly extending TypedEventEmitter, not using the one provided by SharedObject... maybe it should be receiving the emitter in its constructor so Directory can give it its own?