Added printed eslint configs for recommended and strict (#10401)
Printed ESLint configs are helpful to answer the question, "what rules apply to a file?" Storing a printed config in the build makes it easy to tell when a lint rule is changed, and track those changes over time. This PR adds a printed config for the recommended and strict configs.

When I added the printed configs, I was planning to merge in a massive new lint config that would have removed the separate configs completely. So I took a shortcut and didn't print configs for eerything. I should have.

The recommended and strict configs may be merged or removed in the future, but for now it's better to be consistent and have an easy way to answer the question, "is that rule enabled in the strict config?"