Stop using concrete protocol handler class (#10655)
Propagating the refactoring done here #10577 to the client. This slightly simplifies the container op processing and replaces the uses of the protocol handler concrete class with the new interface. We need a better protocol handler creation story in the container. This will be the next step for this work item and will allow the loader to accept a different protocol handler implementation.