Preserve telemetry props when normalizing or wrapping a LoggingError (#10657)
Background: Sometimes we throw a `LoggingError` which has support for telemetry properties, but no `errorType`.  Later that error may pass through a codepath that calls `normalizeError` on it (taking an `unknown` error and returning an error object with `errorType` that supports telemetry properties).

Problem: If you call `normalizeError` on a `LoggingError`, a new error object will be created (since it didn't have `errorType`), but the telemetry properties will not be copied over.