Change Local Reference Event Emitter to Callback (#10677)
Change Local Reference Event Emitter to Callback
The primary impetus for this change is in terms of memory usage. Event emitter is a large class, and making all local references, regardless of usage, an even emitter will result in lots of unnecessary memory being used. this change simplifies the design and uses an optional callback, so when not needed the callbacks will not even be defined, and checking if they are defined or not is a done via a very cheap null coalescing operator.