policy-check fixes (#10696)
In #10583, we ported a change from main into the 1.0 release branch. However, that change contained untagged asserts. The code has since been removed in main (#10583), so it's "safe" to tag it in the release branch, even though that is not what we would typically do. Usually we would cherry-pick the main branch commit that tags the assert. Since the code is removed in main, we're tagging it here.