skip sequence number check when loading with stashed ops (#10704)
There is a check in runtime that ensures the sequence number from the metadata blob in the snapshot from which runtime was loaded matches the initial sequence number given to DeltaManager from which to begin processing new ops. When stashing ops, we save the snapshot we originally loaded from so the metadata blob will be the same, but we supply DeltaManager with the sequence number when the container was closed. Therefore, this check is not relevant when we are loading with stashed ops. The reason this error was not encountered before is because the metadata blob was not present in the saved snapshot in tests, so the check was skipped. I believe it is not included in the summary created when attaching a detached container because it is generated by the runtime.