skip flaky stashed ops test (#10728)
"works with summary while offline" in stashedOps.spec.ts is intermittently hitting a check in SummaryCollection that sends an error event when it sees a summaryAck op without seeing the summary. This is due to SummaryCollection directly listeneing to DeltaManager "op" event, which means it does not see ops replayed during the "apply stashed ops" stage of load.