Encapsulate LocalReference and bi-directionally link to linked list (#10753)
This change is preparation of moving the merge-tree's undo redo reference implementation to leverage local references and is a follow up to moving the internals of local reference collection to a linked this. When we transition undo redo to local references, we will greatly increase the number of add and removes done on local reference, as every undone remove will need to shift the references for any other removed segment occupying the same position in the tree.

To support the above we are now storing a pointer from the position's list in the reference itself. To maintain consistency of this pointer I've limited the code which can access it and fully encapsulated the LocalReference class so it is now only used by the LocalReferenceCollection. Additionally, I've encapulated the other properties of LocalReference that are related to linking to a segment. This will ensure more consistent handling of segment, and offset across all usages.

AB#146
#1009